completed backend type integration
