+++ backend/controllers/k8sController.ts
@@ -1,4 +1,3 @@
-const db = require("../models/alertModel");
 const k8s = require('@kubernetes/client-node');
 
 const kc = new k8s.KubeConfig();
@@ -6,9 +5,13 @@ kc.loadFromDefault();
 
 const k8sApi = kc.makeApiClient(k8s.CoreV1Api);
 
-const k8sController = {};
+import { Request, Response, NextFunction } from 'express';
+import { k8scontroller } from '../../types.ts';
 
-k8sController.getPods = async (req, res, next) => {
+const k8sController: k8scontroller = {getPods: ()=>{}, getNodes: ()=>{}, getServices: ()=>{}};
+
+//gets all pod information from the cluster and passes it into the response locals object
+k8sController.getPods = async (req: Request, res: Response, next: NextFunction) => {
   try {
     const response = await k8sApi.listPodForAllNamespaces();
     res.locals.pods = response.response.body.items;
@@ -18,7 +21,8 @@ k8sController.getPods = async (req, res, next) => {
   }
 }
 
-k8sController.getNodes = async (req, res, next) => {
+//gets all node information from the cluster and passes it into the response locals object
+k8sController.getNodes = async (req: Request, res: Response, next: NextFunction) => {
   console.log('getting nodes');
   try {
     const response = await k8sApi.listNode();
@@ -30,7 +34,8 @@ k8sController.getNodes = async (req, res, next) => {
   }
 }
 
-k8sController.getServices = async (req, res, next) => {
+//gets all service information from the cluster and passes it into the response locals object
+k8sController.getServices = async (req: Request, res: Response, next: NextFunction) => {
   try {
     const response = await k8sApi.listServiceForAllNamespaces();
     res.locals.services = response.response.body.items;
