+++ backend/server.ts
@@ -12,8 +12,8 @@ kc.loadFromDefault();
 const k8sApi = kc.makeApiClient(k8s.CoreV1Api);
 // const mysql = require('mysql');
 
-const alertRouter = require('./router/alertRouter');
-const clusterRouter = require('./router/clusterRouter');
+const alertRouter = require('./router/alertRouter.ts');
+const clusterRouter = require('./router/clusterRouter.ts');
 
 app.use(cors());
 app.use(express.json());
@@ -21,15 +21,6 @@ app.use(express.urlencoded({ extended: true }));
 
 app.use('/alert', alertRouter);
 app.use('/cluster', clusterRouter);
-
-// app.get('/pods', async (req, res) => {
-//   try {
-//     const response = await k8sApi.listPodForAllNamespaces();
-//     res.json(response.body);
-//   } catch (err) {
-//     res.status(500).send(err.message)
-//   }
-// })
 
 app.use((req: Request, res: Response) =>
     res.status(404).send("This is not the page you are looking for...")
