+++ tsconfig.json
@@ -13,6 +13,7 @@
       "moduleResolution": "Node",
       "resolveJsonModule": true,
       "isolatedModules": true,
+      "allowImportingTsExtensions": true,
       "noEmit": true,
       "jsx": "react-jsx"
     },
